Resolve #8 - Implement Testing Routine Implements a testing routine into the Configure object and window. Along with new UI elements.
Keyboard commands are toggled off if testing is started,
and the configuration UI is made unusable until testing ends.

Testing can be stopped by the user clicking the "STOP TEST!" button.
Test parameters are added to the log file if the file is written for testing.

Logging now writes the current input translation and rotation speeds along with the other 4 bits of data.