Chartdisplay WIP Implemented file scanning at startup and selection.
Implemented the basis for rudimentary data selection.
RedrawLines() should edit the visibility of different lines, whenever the data selection is updated. It should call setVisibility(bool) for the different line series, with the argument being tied to the whitelist array.
It's currently a mess of two arrays to make sense of user input.
It may also have issues with accessing non-existent indeces, which causes crashing. Calling chart_.removeSeries() (I think that's the helper), or or removing series invidiually, completely destroys the series object. Even if it's stored elsewhere.

Also has an apparent memory leak (?) with closing (most likely due to bad array deconstruction) and reloading of arrays. The latter may be amended by calling removeSeries() before adding new ones to the chart_ object.